Separate groups into groups overview and groups
